# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: РемонтPRO API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trainings:
    get:
      operationId: getTrainings
      tags: [trainings]
      summary: List trainings
      parameters:
        - name: managerId
          in: query
          schema:
            type: ["integer", "null"]
        - name: difficulty
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of trainings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Training"
    post:
      operationId: createTraining
      tags: [trainings]
      summary: Create a new training
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Training"
  /trainings/{id}:
    get:
      operationId: getTraining
      tags: [trainings]
      summary: Get training by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Training detail with messages and feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDetail"
        "404":
          description: Not found
  /trainings/{id}/message:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a manager message and get AI client reply + coach feedback
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIResponse"
        "400":
          description: Bad request
        "404":
          description: Not found
  /trainings/{id}/finish:
    post:
      operationId: finishTraining
      tags: [trainings]
      summary: Finish training and get final evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Final report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinalReport"
        "400":
          description: Bad request
        "404":
          description: Not found
...
components:
  schemas:
...
    Training:
      type: object
      properties:
        id:
          type: integer
        manager_id:
          type: integer
        manager_name:
          type: string
        difficulty:
          type: string
        property_type:
          type: string
        status:
          type: string
        started_at:
          type: string
        completed_at:
          type: ["string", "null"]
        overall_score:
          type: ["number", "null"]
      required:
        - id
        - manager_id
        - manager_name
        - difficulty
        - property_type
        - status
        - started_at
    TrainingInput:
      type: object
      properties:
        manager_id:
          type: integer
        difficulty:
          type: string
        property_type:
          type: string
      required:
        - manager_id
        - difficulty
        - property_type
    TrainingDetail:
      type: object
      properties:
        id:
          type: integer
        manager_id:
          type: integer
        manager_name:
          type: string
        difficulty:
          type: string
        property_type:
          type: string
        status:
          type: string
        started_at:
          type: string
        completed_at:
          type: ["string", "null"]
        overall_score:
          type: ["number", "null"]
        client_profile:
          type: ["object", "null"]
          additionalProperties: true
        final_report:
          type: ["object", "null"]
          additionalProperties: true
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        feedbacks:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
      required:
        - id
        - manager_id
        - manager_name
        - difficulty
        - property_type
        - status
        - started_at
        - messages
        - feedbacks
    Message:
      type: object
      properties:
        id:
          type: integer
        training_id:
          type: integer
        role:
          type: string
        content:
          type: string
        created_at:
          type: string
      required:
        - id
        - training_id
        - role
        - content
        - created_at
    Feedback:
      type: object
      properties:
        id:
          type: integer
        training_id:
          type: integer
        manager_message_id:
          type: integer
        feedback_json:
          type: object
          additionalProperties: true
        created_at:
          type: string
      required:
        - id
        - training_id
        - manager_message_id
        - feedback_json
        - created_at
    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 2000
      required:
        - content
    AIResponse:
      type: object
      properties:
        client_reply:
          type: string
        coach_feedback:
          type: object
          properties:
            strengths:
              type: string
            mistakes:
              type: string
            next_step:
              type: string
            recommendation:
              type: string
            example_phrase:
              type: string
          required:
            - strengths
            - mistakes
            - next_step
            - recommendation
            - example_phrase
        manager_message_id:
          type: integer
        client_message_id:
          type: integer
        feedback_id:
          type: integer
      required:
        - client_reply
        - coach_feedback
        - manager_message_id
        - client_message_id
        - feedback_id
    FinalReport:
      type: object
      properties:
        overall_score:
          type: number
        scores:
          type: object
          additionalProperties: true
        strengths:
          type: array
          items:
            type: string
        problems:
          type: array
          items:
            type: string
        recommendations:
          type: array
          items:
            type: string
        ideal_strategy:
          type: string
        client_outcome:
          type: string
      required:
        - overall_score
        - scores
        - strengths
        - problems
        - recommendations
        - ideal_strategy
        - client_outcome
...