# lib/api-zod/src/generated/types/aIResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * РемонтPRO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AIResponseCoachFeedback } from './aIResponseCoachFeedback';

export interface AIResponse {
  client_reply: string;
  coach_feedback: AIResponseCoachFeedback;
  manager_message_id: number;
  client_message_id: number;
  feedback_id: number;
}