# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * РемонтPRO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackFeedbackJson } from './feedbackFeedbackJson';

export interface Feedback {
  id: number;
  training_id: number;
  manager_message_id: number;
  feedback_json: FeedbackFeedbackJson;
  created_at: string;
}