# lib/api-zod/src/generated/types/trainingDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * РемонтPRO API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Feedback } from './feedback';
import type { Message } from './message';
import type { TrainingDetailClientProfile } from './trainingDetailClientProfile';
import type { TrainingDetailFinalReport } from './trainingDetailFinalReport';

export interface TrainingDetail {
  id: number;
  manager_id: number;
  manager_name: string;
  difficulty: string;
  property_type: string;
  status: string;
  started_at: string;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  overall_score?: number | null;
  /** @nullable */
  client_profile?: TrainingDetailClientProfile;
  /** @nullable */
  final_report?: TrainingDetailFinalReport;
  messages: Message[];
  feedbacks: Feedback[];
}